multilenguage

+++ public/js/app/app.js
@@ -1,4 +1,3 @@
-
 //const API = "http://smi.alianzacacaoperu.org/api/";
 //const API = "http://localhost:8000/api/";
 
@@ -13,23 +12,23 @@ const KEY_AUTHENTICATED_USER = "auth";
 const URL_HOME = "main";
 const URL_LOGIN = "login";
 
-function authenticatedUser(){    
-    return JSON.parse(localStorage.getItem(KEY_AUTHENTICATED_USER));
+function authenticatedUser() {
+  return JSON.parse(localStorage.getItem(KEY_AUTHENTICATED_USER));
 }
 
-function saveUserCredentials($credentials){
-    localStorage.setItem(KEY_AUTHENTICATED_USER, JSON.stringify($credentials));
+function saveUserCredentials($credentials) {
+  console.log($credentials);
+  localStorage.setItem(KEY_AUTHENTICATED_USER, JSON.stringify($credentials));
 }
 
-function clearCredentials(){
-    if(localStorage.length > 0)
-        localStorage.clear();
+function clearCredentials() {
+  if (localStorage.length > 0) localStorage.clear();
 }
 
-function showLoading(){
-    $('#loading').modal('show');
+function showLoading() {
+  $("#loading").modal("show");
 }
 
-function hideLoading(){
-    $('#loading').modal('hide');
+function hideLoading() {
+  $("#loading").modal("hide");
 }
